Match each record against all the others
This deals with the following case:
Suppose we have a group G. It is made up of records A and B. Record A
matched records B and C above the threshold, but it matched B better, so
that became the group. When we go to match record C, it matches A better
than any other record. If we don't check every record against all the
others, we'll miss this and potentially put C with a worse match.